+++ src/MetaheuristicsPlatform/Algorithms/BiogeographyBasedOptimization/BiogeographyBasedOptimizationReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.BiogeographyBasedOptimization;
+
+public static class BiogeographyBasedOptimizationReferences
+{
+    public static ScientificReference Simon2008 { get; } =
+        new(
+            "Simon",
+            2008,
+            "Biogeography-Based Optimization",
+            "IEEE Transactions on Evolutionary Computation 12(6), 702-713",
+            "10.1109/TEVC.2008.919004");
+}